Fixed bugs and renamed Tmx to Map
