+++ src/Game.hpp
@@ -1,5 +1,4 @@
 #pragma once
-#include <SFML/Graphics.hpp>
 #include <TGUI/TGUI.hpp>
 #include <memory>
 #include <string>
