+++ src/ResourceHandler.cpp
@@ -1,7 +1,5 @@
 #include "ResourceHandler.hpp"
 
-std::unordered_map<std::string, boost::any> ResourceHandler::_cache;
-
 void ResourceHandler::remove(const std::string& key) {
     if (exists(key))
         _cache.erase(key);
