+++ src/Script.hpp
@@ -1,18 +1,16 @@
 #pragma once
 #include <pybind11/eval.h>
 namespace py = pybind11;
-#include <functional>
 #include <string>
 
 class Script final {
 public:
     explicit Script(const std::string& filename);
-
     template <class... Args>
-    inline decltype(auto) operator()(const std::string& name, Args&&... args) {
-        return _local[name.c_str()].cast<py::function>()(std::forward<Args>(args)...);
-    }
+    decltype(auto) operator()(const std::string& name, Args&&... args);
 
 private:
     py::dict _local;
 };
+
+#include "Script.tpp"
